rename infobox to MiningStatistic
