+++ ui/common/MiningStatistic.tsx
