+++ ui/common/index.ts
@@ -7,5 +7,5 @@ export { default as Tips } from './Tips'
 export { default as Tip } from './Tip'
 export { default as Input } from './Input'
 export { default as TextImage } from './TextImageLayout'
-export { default as InfoBox } from './InfoBox'
+export { default as MiningStatistic } from './MiningStatistic'
 export { default as ProgressBar } from './ProgressBar'
